hsu_dma_pci: disable interrupt
On Intel Tangier B0 and Anniedale the interrupt line, disregarding to have
different numbers, is shared between HSU DMA and UART IPs. In that case
IRQ handler is called in UART driver only clearing the interrupt source.

But as the interrupt is level triggered an interrupt storm happens on the
HSU DMA until the interrupt is cleared in the UART driver. To prevent this,
don't enable the HSU DMA interrupt at all.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>